removed devDungeon packages | renamed devdungeon level classes
