+++ devDungeon/src/level/devlevel/BossLevel.java
@@ -3,11 +3,11 @@ package level.devlevel;
 import components.MagicShieldComponent;
 import contrib.components.AIComponent;
 import contrib.components.InventoryComponent;
-import contrib.devDungeon.level.DevDungeonLevel;
 import contrib.entities.AIFactory;
 import contrib.entities.MiscFactory;
 import contrib.item.HealthPotionType;
 import contrib.item.concreteItem.ItemPotionHealth;
+import contrib.level.DevDungeonLevel;
 import contrib.systems.HealthSystem;
 import contrib.utils.EntityUtils;
 import contrib.utils.components.ai.fight.RangeAI;
