+++ devDungeon/src/level/devlevel/IllusionRiddleLevel.java
@@ -5,10 +5,10 @@ import contrib.components.AIComponent;
 import contrib.components.HealthComponent;
 import contrib.components.InteractionComponent;
 import contrib.components.InventoryComponent;
-import contrib.devDungeon.level.DevDungeonLevel;
 import contrib.entities.MiscFactory;
 import contrib.item.HealthPotionType;
 import contrib.item.concreteItem.ItemPotionHealth;
+import contrib.level.DevDungeonLevel;
 import contrib.utils.EntityUtils;
 import contrib.utils.components.ai.fight.RangeAI;
 import core.Entity;
