+++ devDungeon/src/level/devlevel/TutorialLevel.java
@@ -4,11 +4,11 @@ import com.badlogic.gdx.Input;
 import contrib.components.HealthComponent;
 import contrib.components.InventoryComponent;
 import contrib.configuration.KeyboardConfig;
-import contrib.devDungeon.level.DevDungeonLevel;
 import contrib.entities.MiscFactory;
 import contrib.hud.DialogUtils;
 import contrib.item.HealthPotionType;
 import contrib.item.concreteItem.ItemPotionHealth;
+import contrib.level.DevDungeonLevel;
 import contrib.utils.EntityUtils;
 import contrib.utils.components.skill.SkillTools;
 import core.Entity;
